requirements.txt
telethon==1.33.1
discord.py-self==2.0.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
loguru==0.7.2
pyaes==1.6.1
rsa==4.9
